Don't convert nil slices into empty, non-nil slices
If a slice is nil, the json package marshals it into `null`. If it's
empty, but non-nil, json marshals it into `[]`. Similarly, the json
package unmarshals `null` into `[]interface{}(nil)`, but `[]` into
`[]interface{}{}`, which are unequal.

As a result, the previous functionality created a weird situation where
the object that was unmarshaled after a round-trip to rethink was
unequal to the original object. That behavior is unusual as we should
expect that the objects would be equal.

This commit ensures that nil slices retain their nil-ness when encoded
by gorethink, which ensures that json encodes them as `null` and then
decodes them as `[]inteface{}(nil)`, which is the same value they had
before encoding.