Python 3.8 now uses ast.Constant for all constants.  This script must now be run with 3.8 or newer.
It's now possible to disable specific lines from the YAML test cases.
I encountered a number of edge cases that were too troublesome or
time consuming to fix.  The impact should be minor since the other
cases in the affected files are still included.  As some
compensation, some of the previously excluded test files are now
partially included.